Fixed new item order IDs not populating
- Newly created items from a newly created order now populate order ID correctly
- Changed a few of the ID types to uint to better reflect how gorm stores it

+++ pkg/repository/item_repository.go
@@ -12,11 +12,11 @@ type Item interface {
 	// Exists determines if an item with the supplied id exists.
 	Exists(id int) (bool, error)
 	// GetByID finds and returns an individual item with the supplied id. Returns nil on error.
-	GetByID(id int) (*models.Item, error)
+	GetByID(id uint) (*models.Item, error)
 	// GetByOrderID returns an array of all the items in the repository with the supplied order id. Returns nil on error.
-	GetByOrderID(id int) ([]*models.Item, error)
+	GetByOrderID(id uint) ([]*models.Item, error)
 	// Create creates a new item and places it in the repository. Returns the ID of the newly created item, -1 on error.
-	Create(i *models.Item) (int64, error)
+	Create(i *models.Item) (uint, error)
 	// Update updates an existing item in the repository. Returns nil on error.
 	Update(i *models.Item, fields []string) (*models.Item, error)
 	// Delete removes an item with the supplied id from the repository. Returns true on success, false on error.
